requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
python-dotenv==1.0.0
supabase==1.0.3